:art: serial

+++ src/STSServoDriver.hpp
@@ -100,7 +100,7 @@ public:
   /// \param serialPort Serial port, default is Serial
   /// \param baudRate Baud rate, default 1Mbps
   /// \returns  True on success (at least one servo responds to ping)
-  bool init(gpio_num_t const& dirPin, SerialPort* serialPort = nullptr, long const& baudRate = 1000000);
+  bool init(gpio_num_t const& dirPin, SerialPort* serialPort = nullptr, int const& baudRate = 1000000);
 
   /// \brief Initialize the servo driver without direction pin.
   /// when you use an interface board including FE-URT-1, direction pin is not required.
@@ -109,7 +109,7 @@ public:
   /// \param serialPort Serial port, default is Serial
   /// \param baudRate Baud rate, default 1Mbps
   /// \returns  True on success (at least one servo responds to ping)
-  bool init(SerialPort* serialPort = nullptr, long const& baudRate = 1000000);
+  bool init(SerialPort* serialPort = nullptr, int const& baudRate = 1000000);
 
   /// \brief Ping servo
   /// \param[in] servoId ID of the servo
